Cleaned up and exported the modal component

+++ packages/admin/src/components/DeleteAllShareLinksModal/DeleteAllShareLinksModal.tsx
@@ -1,5 +1,5 @@
 import type { Accessor, Component } from "solid-js";
-import { ConfirmationModal } from "@/components/ConfirmationModal/ConfirmationModal";
+import { Modal } from "@/components/Modal/Modal";
 import api from "@/services/api";
 import T from "@/translations";
 
@@ -25,30 +25,23 @@ const DeleteAllShareLinksModal: Component<DeleteAllShareLinksProps> = (
 	// ------------------------------
 	// Render
 	return (
-		<ConfirmationModal
-			state={{
-				open: props.state.open,
-				setOpen: props.state.setOpen,
-				isLoading: deleteAllShareLinks.action.isPending,
-				isError: deleteAllShareLinks.action.isError,
+		<Modal.Confirm
+			open={props.state.open}
+			onOpenChange={props.state.setOpen}
+			title={T()("modals.common.delete.all.share.links.title")}
+			description={T()("modals.common.delete.all.share.links.description")}
+			loading={deleteAllShareLinks.action.isPending}
+			error={deleteAllShareLinks.errors()?.message}
+			onConfirm={() => {
+				const id = props.id();
+				if (!id) return console.error("No id provided");
+				deleteAllShareLinks.action.mutate({
+					mediaId: id,
+				});
 			}}
-			copy={{
-				title: T()("modals.common.delete.all.share.links.title"),
-				description: T()("modals.common.delete.all.share.links.description"),
-				error: deleteAllShareLinks.errors()?.message,
-			}}
-			callbacks={{
-				onConfirm: () => {
-					const id = props.id();
-					if (!id) return console.error("No id provided");
-					deleteAllShareLinks.action.mutate({
-						mediaId: id,
-					});
-				},
-				onCancel: () => {
-					props.state.setOpen(false);
-					deleteAllShareLinks.reset();
-				},
+			onCancel={() => {
+				props.state.setOpen(false);
+				deleteAllShareLinks.reset();
 			}}
 		/>
 	);
